stage-3/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

sentence-transformers
numpy<2.0
faiss-gpu

llama-cpp-python
huggingface-hub